clean up and move to payment and orders

+++ src/address/address.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CreateAddressDto } from './dto/create-address.dto';
 import { UpdateAddressDto } from './dto/update-address.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -8,24 +13,38 @@ import { buildPaginatedResponse } from 'src/common/pagination.response';
 import { CurrentUser } from 'src/utility/decorators/CurrentUser.decorator';
 import { UserEntity } from 'src/users/entities/user.entity';
 import { AddressType } from 'src/utility/enums/address.enum';
+import { UsersService } from 'src/users/users.service';
 
 @Injectable()
 export class AddressService {
   constructor(
     @InjectRepository(AddressEntity)
     private readonly addressRepository: Repository<AddressEntity>,
+
+    private readonly usersService: UsersService,
   ) {}
 
-  async create(createAddressDto: CreateAddressDto) {
+  async create(createAddressDto: CreateAddressDto, currentUser: UserEntity) {
+    const user = await this.usersService.findById(currentUser.user_id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
     const address = this.addressRepository.create(createAddressDto);
 
-    return await this.addressRepository.save(address);
+    await this.addressRepository.save({ ...address, user });
+
+    return {
+      statusCode: 201,
+      message: 'Address created successfully',
+      data: address,
+    };
   }
 
   async findAll(page: number = 1, limit: number = 10) {
     const [data, totalCount] = await this.addressRepository.findAndCount({
       skip: (page - 1) * limit,
       take: limit,
+      relations: ['user'],
       order: {
         created_at: 'DESC',
       },
@@ -41,7 +60,17 @@ export class AddressService {
   }
 
   async findOne(id: string): Promise<AddressEntity> {
-    const address = await this.addressRepository.findOneBy({ address_id: id });
+    const address = await this.addressRepository.findOne({
+      where: { address_id: id },
+      relations: ['user'],
+      select: {
+        user: {
+          user_id: true,
+          name: true,
+          email: true,
+        },
+      },
+    });
 
     if (!address) {
       throw new NotFoundException('Address not found');
@@ -49,26 +78,97 @@ export class AddressService {
     return address;
   }
 
-  async findByUser(user: UserEntity) {
+  async findByUser(id: string) {
+    const user = await this.usersService.findById(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    const address = await this.addressRepository.find({
+      relations: ['user'],
+      where: { user: { user_id: id } },
+      select: {
+        user: {
+          user_id: true,
+          name: true,
+          email: true,
+        },
+      },
+    });
+
+    return {
+      statusCode: 200,
+      message: 'Address fetched successfully',
+      data: address,
+    };
+  }
+
+  async findByLoggedInUser(user: UserEntity) {
     return await this.addressRepository.find({
       where: { user: { user_id: user.user_id } },
     });
   }
 
-  async update(id: string, updateAddressDto: UpdateAddressDto) {
-    return await this.addressRepository.update(id, updateAddressDto);
+  async update(
+    id: string,
+    updateAddressDto: UpdateAddressDto,
+    currentUser: UserEntity,
+  ) {
+    const address = await this.findOne(id);
+    if (address.user.user_id !== currentUser.user_id) {
+      throw new UnauthorizedException('You can only update your own address');
+    }
+
+    const payload = {
+      ...address,
+      ...updateAddressDto,
+    };
+    await this.addressRepository.update(id, payload);
+    return {
+      statusCode: 200,
+      message: 'Address updated successfully',
+      data: payload,
+    };
   }
-  async setAsBilling(id: string) {
-    return await this.addressRepository.update(id, {
+  async setAsBilling(id: string, currentUser: UserEntity) {
+    const address = await this.findOne(id);
+    if (address.user.user_id !== currentUser.user_id) {
+      throw new UnauthorizedException('Access denied');
+    }
+
+    await this.addressRepository.update(id, {
       type: AddressType.BILLING,
     });
+
+    return {
+      statusCode: HttpStatus.OK,
+      message: 'Address updated successfully',
+      data: await this.findOne(id),
+    };
   }
 
-  async setAsDefault(id: string) {
-    return await this.addressRepository.update(id, { is_default: true });
+  async setAsDefault(id: string, currentUser: UserEntity) {
+    const address = await this.findOne(id);
+    if (address.user.user_id !== currentUser.user_id) {
+      throw new UnauthorizedException('Access denied');
+    }
+    await this.addressRepository.update(id, {
+      is_default: !address.is_default,
+    });
+    return {
+      statusCode: HttpStatus.OK,
+      message: 'Address updated successfully',
+      data: await this.findOne(id),
+    };
   }
 
-  async remove(id: string) {
+  async remove(id: string, currentUser: UserEntity) {
+    const address = await this.findOne(id);
+
+    if (address.user.user_id !== currentUser.user_id) {
+      throw new UnauthorizedException('You can only delete your own address');
+    }
     return await this.addressRepository.delete({ address_id: id });
   }
 }
